# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HiGun API",
    "description": "HiGun is a MetaID-based UTXO indexer supporting FT (Fungible Token) and NFT (Non-Fungible Token) chain data queries.",
    "version": "1.0.0",
    "contact": {
      "name": "HiGun Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Default server"
    }
  ],
  "paths": {
...
    "/ft/blocks/reindex": {
      "get": {
        "summary": "Reindex Blocks",
        "description": "Reindex blocks in specified range",
        "operationId": "reindexBlocks",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "description": "Start block height",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "End block height",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
...
  },
...
}